feat(build,msan): MSAN sanitizer support + dual-mode thirdparty (source/prebuilt)
This squashes 28 commits that together deliver:

1. **Thirdparty CMake-ification with dual-mode build**
   - thirdparty/CMakeLists.txt + thirdparty/cmake/<lib>.cmake per library:
     replaces autoconf+make in build-thirdparty.sh with CMake targets that
     can be added via add_subdirectory(thirdparty) from BE's main CMake.
   - USE_CONTRIB_SOURCE option (default ON): switches between
     - source mode: add_subdirectory(thirdparty) — flags (incl. sanitizer
       flags) propagate to every thirdparty .o
     - prebuilt mode: include(be/cmake/thirdparty.cmake) — IMPORTED .a from
       \$DORIS_THIRDPARTY/installed, matches legacy doris/master flow
   - Sanitizer builds (ASAN/TSAN/MSAN/LSAN/UBSAN) force source mode (prebuilt
     .a can't be sanitizer-instrumented). --use-prebuilt with a sanitizer is
     a hard cmake error.
   - build.sh exposes --use-source / --use-prebuilt flags and auto-runs
     download-thirdparty.sh on a fresh-clone source build, or
     download-prebuild-thirdparty.sh + tar -Jxf on a fresh-clone
     --use-prebuilt build. Fresh git clone of doris just works in either
     mode with one command.

2. **MSAN sanitizer build framework** (new build type)
   - be/CMakeLists.txt MSAN section: -fsanitize=memory
     -fsanitize-recover=memory -fsanitize-memory-track-origins=2
     -fsanitize-memory-use-after-dtor, force USE_LIBCPP, force USE_JEMALLOC
     OFF, disable libunwind / glibc-compatibility (incompatible).
   - thirdparty/libcxx-msan-cmake/ + thirdparty/libcxxabi-msan-cmake/:
     LLVM runtimes rebuilt with LLVM_USE_SANITIZER=MemoryWithOrigins,
     statically linked into BE under MSAN.
   - conf/msan_ignorelist.txt: 5 patterns, all libunwind (CPU register
     reads cause MSAN abort to recurse into libunwind itself — physical
     lower bound, matches LLVM upstream consensus). Other libraries are
     fully instrumented.
   - be/src/common/memory_sanitizer.h: centralized no-op header so non-MSAN
     builds compile unchanged when call sites use __msan_unpoison etc.
   - bin/start_be.sh: MSAN_OPTIONS (halt_on_error=0, abort_on_error=0,
     poison_in_dtor=1, log_path) + ulimit so BE survives a warning storm.
   - be/CMakeLists.txt msan_stubs target provides HeapProfilerStart/Stop/
     GetHeapProfile stubs + __msan_default_options to handle init-time
     false positives from uninstrumented static constructors.

3. **fmt 7.1.3 → 8.1.1 upgrade** (required to drop noisy ignorelist)
   - thirdparty/vars.sh + thirdparty/cmake/fmt.cmake bump version.
   - be/cmake/thirdparty.cmake (prebuilt mode) rebuilds fmt 8 from source
     anyway because the prebuilt tarball still ships fmt 7.1.3, whose v7::*
     symbols would mismatch BE's v8 includes.
   - 13 BE .cpp/.h adapt to fmt 8 idioms (fmt::runtime() for runtime format
     strings, std::back_inserter for memory_buffer, explicit enum casts).

4. **Library instrumentation (Wave 11-20)**
   - Sequentially removed every removable thirdparty from msan_ignorelist:
     mysql-server / unixODBC (header-only), header-only SIMD libs
     (AvxToNeon / sse2neon / xsimd / libdivide), bitshuffle / streamvbyte,
     simdjson, hyperscan, libbacktrace, faiss, jemalloc (skip compile under
     MSAN; sanitizer runtime allocator takes over), glibc-compatibility
     (dead under MSAN). Plus historical Wave 1-10 (compression, network,
     crypto, serialization, storage engines, cloud SDKs, misc).
   - hadoop_hdfs.cmake forwards parent CMAKE_C_FLAGS into its sub-cmake so
     libhdfs is sanitizer-instrumented (was silently uninstrumented
     before).
   - thirdparty/patches/doris-thirdparty-hadoop-3.3.6.6-for-doris-msan.patch:
     11 HDFS_MSAN_UNPOISON calls at JNI Get*Region / GetStringUTFChars
     boundaries inside libhdfs (JVM-written buffers are invisible to MSAN
     otherwise). download-thirdparty.sh applies the patch automatically.

5. **Real Doris bugs surfaced by MSAN**
   - be/src/runtime/runtime_state.h: init _load_job_id = -1.
   - be/src/storage/field.h: init _precision = 0, _scale = 0.
   - be/src/core/block/block.h: init IteratorRowRef row_pos / is_same.
   - be/src/exprs/vexpr.h: init _opcode / _node_type.
   - be/src/core/pod_array.h: __msan_unpoison(c_end_of_storage, pad_right)
     so the over-allocated tail is not flagged as uninit.
   - be/src/common/multi_version.h: replace std::atomic<shared_ptr<T>> with
     a wrapper to compile against doris's libc++.

6. **UBSAN build fixes** (independent — UBSAN was previously broken in
   non-MSAN configurations too)
   - Clear UBSAN_LIBS on clang path (the old -rtlib=compiler-rt dropped
     libgcc_s and broke _Unwind_* resolution; sanitizer runtime is
     auto-linked by -fsanitize=undefined).
   - pprof_actions.cpp: add UNDEFINED_BEHAVIOR_SANITIZER to the ifdef
     guards that stub HeapProfilerStart/Stop/CPU profiling (UBSAN was
     missing from those guards, so it tried to call gperftools heap
     profiler APIs that aren't linked under sanitizer builds).
   - be/CMakeLists.txt: aws-c-common generated config.h include path
     pointed at the wrong directory; fix lets ASAN/TSAN/UBSAN/LSAN fresh
     --clean builds succeed first-pass without a retry.

7. **Verified build matrix** (all from fresh state, build_*/ removed):
   - Release × source, Release × prebuilt
   - Debug × source, Debug × prebuilt
   - MSAN × source (forced)
   - ASAN × source, TSAN × source, LSAN × source, UBSAN × source
   - sanitizer + --use-prebuilt → cmake FATAL_ERROR (by design)

8. **Documentation**
   - thirdparty/README.md: dual-mode operation, auto-download flow,
     reconcile table for source-vs-prebuilt target name mismatches,
     sub-cmake forward-flags gotcha.

+++ be/src/util/s3_util.h
@@ -112,8 +112,8 @@ struct S3ClientConf {
                 "request_timeout_ms={}, connect_timeout_ms={}, use_virtual_addressing={}, "
                 "cred_provider_type={},role_arn={}, external_id={}",
                 hide_access_key(ak), token, endpoint, region, bucket, max_connections,
-                request_timeout_ms, connect_timeout_ms, use_virtual_addressing, cred_provider_type,
-                role_arn, external_id);
+                request_timeout_ms, connect_timeout_ms, use_virtual_addressing,
+                static_cast<int>(cred_provider_type), role_arn, external_id);
     }
 };
